made code C++ fully generic for working precision

+++ tests/drivers/factor_node_test.cxx
@@ -26,7 +26,8 @@ int main(int argc, char** argv) {
 
    if (opts.chol) {
 
-      spldlt::tests::factor_node_posdef_test<double>(opts.m, opts.n, opts.nb, opts.ncpu, opts.ngpu, opts.check);
+      // spldlt::tests::factor_node_posdef_test<double>(opts.m, opts.n, opts.nb, opts.ncpu, opts.ngpu, opts.check);
+      spldlt::tests::factor_node_posdef_test<float>(opts.m, opts.n, opts.nb, opts.ncpu, opts.ngpu, opts.check);
    }
    else {
       // factor_node_indef_test<double, 32, false>(0.01, 1e-20, true, false, opts.m, opts.n, opts.nb, opts.ncpu);
